+++ Sources/ZcashLightClientKit/Error/ZcashErrorCode.swift
@@ -327,6 +327,8 @@ public enum ZcashErrorCode: String {
     case rustMigrationLockResidual = "ZRUST0132"
     /// Error from rust layer when calling ZcashRustBackend.unlockMigrationResidual
     case rustMigrationUnlockResidual = "ZRUST0133"
+    /// Error from rust layer when calling ZcashRustBackend.estimateMigrationRuns
+    case rustMigrationEstimateRuns = "ZRUST0134"
     /// SQLite query failed when fetching all accounts from the database.
     case accountDAOGetAll = "ZADAO0001"
     /// Fetched accounts from SQLite but can't decode them.